feat: 일부 코드 typescript로 변환
