+++ package.json
@@ -33,6 +33,7 @@
   },
   "devDependencies": {
     "@shelf/jest-mongodb": "^1.1.3",
+    "@types/express": "^4.17.9",
     "husky": "^3.0.9",
     "jest": "^24.9.0",
     "lint-staged": "^9.4.3",
